DDPB-1247: Using a trait for address to reduce code duplication
